Restructure server/lib/forms into self-contained form folders
- Create cert/, 849b/, 647f/ folders, each with metadata.js,
  generate.js, and index.js barrel
- Move shared files (FormContainer, formComponents, formUtils,
  pdf-forms.css) into shared/
- Move PDF templates/fonts to their respective form folders
- Update getFormMetadata.js to import directly from form barrels
  instead of dynamic dist imports
- Simplify generateFormPdfBuffer — remove cacheBust/componentName
- Update esbuild entry points to new paths
- Update client formRegistry.js and FormPreview.jsx import paths
- Remove cacheBust option from routes/api/forms/index.js
- Delete shared/src/forms.js (metadata now lives in form folders)
