+++ server/esbuild.forms.js
@@ -4,8 +4,10 @@ const watch = process.argv.includes('--watch');
 
 const options = {
   entryPoints: [
-    'lib/forms/jsx/*.jsx',
-    'lib/forms/pdf/*.jsx',
+    'lib/forms/647f/Form647f.jsx',
+    'lib/forms/cert/NarcoticsNotice.jsx',
+    'lib/forms/shared/FormContainer.jsx',
+    'lib/forms/shared/formComponents.jsx',
   ],
   outdir: 'lib/forms/dist',
   entryNames: '[name]',
