+++ server/lib/forms/647f/Form647f.jsx
@@ -1,8 +1,6 @@
 import React from 'react';
-import { z } from 'zod';
-import { FORM_TIMEZONE, formatDateTime24, formatDateOnly, titleCase } from '../formUtils.js';
-import { generateJsxPdf } from '../generate.js';
-import { Header, Row, SectionHeader } from './formComponents.jsx';
+import { FORM_TIMEZONE, formatDateTime24, formatDateOnly, titleCase } from '../shared/formUtils.js';
+import { Header, Row, SectionHeader } from '../shared/formComponents.jsx';
 
 /*
  * Override pdf-forms.css page constraints so this form can grow beyond
@@ -28,130 +26,6 @@ const pageCSS = `
     }
   }
 `;
-
-export const metadata = {
-  canGenerate () {
-    return true; // No preconditions — can be generated at any point in the deflection lifecycle
-  },
-
-  deflectionInclude: {
-    subject: true,
-    incident: {
-      include: {
-        createdBy: {
-          include: {
-            organization: true,
-            unit: true,
-          },
-        },
-        incidentOfficers: {
-          include: {
-            officer: {
-              include: {
-                organization: true,
-                unit: true,
-              },
-            },
-            organization: true,
-            unit: true,
-          },
-        },
-      },
-    },
-    facility: true,
-    createdBy: {
-      include: {
-        organization: true,
-        unit: true,
-      },
-    },
-  },
-
-  dataSchema: z.object({
-    deflectionId: z.number(),
-    subjectLastName: z.string(),
-    subjectFirstName: z.string(),
-    subjectMiddleInitial: z.string(),
-    subjectRace: z.string(),
-    subjectSex: z.string(),
-    subjectDOB: z.string().nullable(),
-    subjectAddress: z.string(),
-    subjectDL: z.string(),
-    subjectLocalId: z.string(),
-    cadNumber: z.string(),
-    arrestedAt: z.string().nullable(),
-    officerName: z.string(),
-    arrestLocation: z.string(),
-    officerUnit: z.string(),
-    officerBadge: z.string(),
-    supervisorBadgeNumber: z.string(),
-    agency: z.string(),
-    charge: z.string(),
-    justification: z.string(),
-    substanceFound: z.boolean(),
-    paraphernaliaFound: z.boolean(),
-    facilityName: z.string(),
-    facilityAddress: z.string(),
-  }),
-
-  transformData (deflection) {
-    const subject = deflection.subject;
-    const incident = deflection.incident;
-    const arrestingOfficerRecord = incident?.incidentOfficers?.find(record => record.role === 'ARRESTING');
-    const officer = arrestingOfficerRecord?.officer || incident?.createdBy || deflection.createdBy;
-
-    const subjectAddress = [subject?.addressLine1, subject?.city, subject?.state]
-      .filter(Boolean)
-      .join(', ');
-
-    const arrestLocation = [incident?.addressLine1, incident?.city, incident?.state]
-      .filter(Boolean)
-      .join(', ');
-
-    const officerName = officer
-      ? `${officer.firstName} ${officer.lastName}`
-      : '';
-    const officerBadge = arrestingOfficerRecord?.badgeNumber || incident?.createdByBadgeNumber || officer?.badgeNumber || '';
-    const officerUnit = arrestingOfficerRecord?.unit?.name || incident?.createdByUnit?.name || officer?.unit?.name || '';
-    const agency = arrestingOfficerRecord?.organization?.name || officer?.organization?.name || '';
-
-    const facility = deflection.facility;
-    const facilityAddress = [facility?.addressLine1, facility?.city, facility?.state, facility?.postalCode]
-      .filter(Boolean)
-      .join(', ');
-
-    return {
-      deflectionId: deflection.id,
-      subjectLastName: subject?.lastName || '',
-      subjectFirstName: subject?.firstName || '',
-      subjectMiddleInitial: subject?.middleInitial || '',
-      subjectRace: subject?.race || '',
-      subjectSex: subject?.sex || '',
-      subjectDOB: subject?.dateOfBirth?.toISOString() || null,
-      subjectAddress,
-      subjectDL: subject?.driverLicense || '',
-      subjectLocalId: subject?.localId || '',
-      cadNumber: incident?.cadNumber || '',
-      arrestedAt: incident?.arrestedAt?.toISOString() || null,
-      officerName,
-      arrestLocation,
-      officerUnit,
-      officerBadge,
-      supervisorBadgeNumber: incident?.supervisorBadgeNumber || '',
-      agency,
-      charge: '647(f) RWS',
-      justification: deflection.behavior || '',
-      substanceFound: deflection.narcoticsSubstance === true,
-      paraphernaliaFound: deflection.narcoticsParaphernalia === true,
-      facilityName: facility?.name || '',
-      facilityAddress,
-    };
-  },
-
-  async generatePdf (formData) {
-    return generateJsxPdf(Form647f, formData, 'SFPD 647(f) Report');
-  },
-};
 
 export default function Form647f ({ data = {} }) {
   const {
