+++ server/lib/forms/849b/index.js
@@ -1,0 +1,2 @@
+export { metadata } from './metadata.js';
+export { generatePdf } from './generate.js';
