+++ server/lib/forms/generate.js
@@ -8,10 +8,6 @@ export async function generateJsxPdf (FormComponent, formData, title) {
 }
 
 // generates the PDF buffer for a single form.
-// opts.cacheBust - if true, appends a timestamp query string to dynamic imports
-//   to bypass Node's module cache (used in development)
-export async function generateFormPdfBuffer (form, formData, user, opts = {}) {
-  const cacheBust = opts.cacheBust ? `?t=${Date.now()}` : '';
-  const { metadata } = await import(`#lib/forms/dist/${form.componentName}.js${cacheBust}`);
-  return Buffer.from(await metadata.generatePdf(formData, user));
+export async function generateFormPdfBuffer (form, formData, user) {
+  return Buffer.from(await form.generatePdf(formData, user));
 }
